pyproject.toml
[tool.poetry]
name = "demo-acoustic-neural-distance"
version = "0.1.0"
description = ""
authors = ["Raoul Buurke <raoul.buurke@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.1"
soundfile = "^0.12.1"
torch = "^2.2.1"
transformers = "^4.38.2"
praatio = "^6.2.0"
librosa = "^0.10.1"
dtw-python = "^1.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
